Fix: client-component separation
